First beta of an emulation application with FdNetDevice
This commit introduces a first draft version of an emulation
application which emulates N vehicles (for the time being, relying
on the same scenario as the one used for the V2I applications with
the network simulated by ns-3).
This application emulates N vehicles, each with its own CA and DEN
basic service (+ BTP and GeoNetworking), and make them send the CAM
messages and receive the DENM messages through a physical interface
(specified with the "interface" option), instead of using any ns-3
simulated model.
This should enable, in the future, hardware-in-the-loop testing and
evaluation.
The application is called, for the time being, "emu-v2x" and it
requires the specified interface to be in promiscuous mode
(sudo ip link set <interface name> promisc on).
'sudo' may be needed to use the underlying FdNetDevice.