New advanced option to set SUMO network namespace
This commit adds a new option for the TraCI Client
("UseNetworkNamespace") which can be used by advanced
users to tell ms-van3t to launch sumo inside a given
Linux network namespace.
The new option can be set on the main() function of the
application with:
sumoClient->SetAttribute("UseNetworkNamespace",
	StringValue(<std::string containing the namespace name>));
This can be useful when the whole ms-van3t (i.e. ./waf run)
is launched inside a namespace with "sudo ip netns exec <name>".
Without this option, as the TraCI communication happens on
localhost, ns-3 would run fine on the network namespace,
but SUMO would not be able to communicate with ns-3, as it
is launched, by default, on the root namespace.
Thus, when ms-van3t is launched inside a namespace, the new
option shall also be specified, setting the same namespace as
the one used to launch ms-van3t.
The v2x-emulator application has also been updated to let the
user set the name of the SUMO network namespace, with
--sumo-netns <namespace name>
(if needed, otherwise SUMO is launched normally)