Enhancements to the gps-tc module
This commit introduces few enhancements to the gps-tc module,
enabling the possibility of using GNSS traces coming from
potentially any kind of device (now the names of the columns
containing the needed data inside the CSV file can be specified
by the user and, if the acceleration is not provided, it can
be assumed as constant and automaticcally computed between
couple of points).
It also introduces an additional bug fix in the gps-tc example:
now, when the GPS Traces of all the nodes are terminated, the
simulation will terminate too, instead of waiting for the whole
simulation time.
It also adds the TransverseMercator_Reverse() function to the
C-port of the TransverseMercator GeographicLib module.