New BSContainer wrapper, improved TraCI API and misc improvements
This commit introduces a new wrapper to the CA and DEN Basic Services,
namely, the Basic Service Container (BSContainer). A BSContainer
object can be used to easily create and configure both ETSI Basic
Services for a vehicle.
It also introduces a new class, called BSMap, as a data structure
to store a set of BSContainers for different vehicles, and easily
retrieve the right container starting from the vehicle ID.
Furthermore, this commit also enhances the TraCI API implementation,
by adding the "highlight" method.
Finally, a few misc improvements are introduced.