TraCI interface improvement
This commit improves the TraCI interface to return the SUMO ID
in both the node creation (m_includeNode) and node shutdown
(m_excludeNode) callbacks. All the examples have been updated
to use the proper callback format in setupNewWifiNode() and
shutdownWifiNode().
This commit introduces also an easy-to-use script to enable the
usage of the v2x-emulator even when --enable-sudo is not specified.
This script is located within the main ns-3-dev directory and it
is called "enable_v2x_emulator.sh".
This commit also includes a new "cleanup" function for the BSContainer,
to be called inside the node shutdown callback, in scenarios where
vehicles can exit the map before the end of the simulation time.