Fix the desynchronization mechanism for both CAMs and CPMs
This commit fixes the desync mechanism for both CAMs and CPMs,
to properly simulate different startup times for the vehicles
networking stacks (as it would happen in a real-world case).
This also avoids, in certain situations, packet losses due to
perfectly synchronized message transmissions.
It also fixes possible crashes that were occurring when vehicles
were exiting the simulation before the end of the selected
simulation time.