Improve the PRRSupervisor
This commit improves the PRRSupervisor to be more resilient to
simulation configuration errors, that could previously lead to
inconsistent PRR values (> 1).
In addition, it is now possible to exclude from the PRR and
latency computation vehicles that send and receive ETSI C-ITS
messages, alongside vehicles that are not connected or send
an interfering traffic (as it was possible until now).