Fix bug in DEN Basic Service and introduce a new example
This commit fixes a small bug in the DEN Basic Service when a Situation container is used.
It also updates README.md, it introduces a new utility function for logging to CSV files (writeDataToCSV) and it introduces a new sample application modelling a DSRC V2V scenario configured with a single .cc file, using the BSContainer ms-van3t facility and with vehicle sending CAMs, except for one vehicle that is sending an interfering traffic to simulate a fairly congested channel.
This example can also be used as a simple application to see the effect of setting different Access Categories in IEEE 802.11p.