Introduction of a full VRU Basic Service (VBS)
This commit introduces for the first time the VRU Awaraness Messages
(VAMs) into ms-van3t, together with a full implementation of the VBS.
It also introduces support for the simulation of VRUs when leveraging
SUMO as mobility simulator, thanks to an extension to the TraCI client.
All the features foreseen by ETSI are available, except for clustering
and interception probability, that are currently being developed.
The simulation of VRUs when using CARLA is also under development.